+++ src/modules/browser/browser_debug.h
@@ -11,4 +11,4 @@
 
 /* `font_scale` should be the view's current browser_view_state_t.font_scale
  * (see browser_render.h) so the printed layout matches what's on screen. */
-void browser_debug__dump(const browser_document_t *doc, int font_scale);
+void browser_debug__dump(const browser_document_t *doc, float font_scale);
